create guest layout
