+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Middleware;
 
+use App\Http\Resources\UserResource;
 use Illuminate\Http\Request;
 use Inertia\Middleware;
 
@@ -22,6 +23,10 @@ class HandleInertiaRequests extends Middleware
      */
     public function share(Request $request): array
     {
-        return array_merge(parent::share($request), []);
+        return array_merge(parent::share($request), [
+            "appName" => config("app.name"),
+            "user" => fn() => $request->user() ? UserResource::make($request->user()) : null,
+            "status" => fn() => $request->session()->get("message"),
+        ]);
     }
 }
