+++ resources/js/Layouts/BaseLayout.vue
@@ -3,16 +3,28 @@
 import type {Page} from '@/Types/Page'
 import Header from '@/components/Common/Header.vue'
 import Footer from '@/components/Common/Footer.vue'
-import BackgroundEffect from '@/components/Home/BackgroundEffect.vue'
+import BackgroundEffect from '@/components/Common/BackgroundEffect.vue'
+import Banner from '@/components/Common/Banner.vue'
+import { usePage } from '@inertiajs/vue3'
+import {ref} from 'vue'
+import {type PageProps} from '@/Types/PageProps'
 
-defineProps<{ pages: Page[] }>()
+const { pages } = defineProps<{ pages: Page[] }>()
+
+const props = usePage().props as PageProps
+const status = ref<string>(props.status ?? '')
+
 </script>
 
 <template>
   <BackgroundEffect />
-  <div class="inset-0 bg-white/60 fixed -z-10" />
+
   <div class="flex flex-col items-center h-full min-h-screen">
-    <Header :pages />
+    <Transition>
+      <Banner v-if="status" :text="status" @click="status = ''" />
+    </Transition>
+
+    <Header :pages :user="props.user" :app-name="props.appName" />
     <slot />
     <Footer />
   </div>
