+++ resources/js/Types/PageProps.d.ts
@@ -1,0 +1,7 @@
+import {type User} from '@/Types/User'
+
+export interface PageProps {
+  appName: string
+  status?: string
+  user?: User
+}
