--- resources/js/components/Common/BackgroundEffect2.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-import BackgroundEffect from '../Home/BackgroundEffect.vue'
-</script>
-
-<template>
-  <div class="absolute opacity-50 -z-50">
-    <BackgroundEffect />
-  </div>
-  <div class="bg-primary-600/10 size-full fixed top-0 left-0 -z-40" />
-</template>
